refactor: split naver map marker lifecycle hooks
